add the polygon netwrok token scripts and contract burn and mint functions
